TensorFlow/Segmentation/VNet/requirements.txt
SimpleITK==1.1.0
requests
googledrivedownloader
tf2onnx
